Added syntax_errors.c
I added a syntax_errors.c file which in there is a function called handle_syntax_errors. It checks if the input has any possible syntax errors which bash gives a message out. Were checking (hopefully) every single syntax error possible. If a syntax error appears we are giving a message out. Calling the function in our parsing.c in the parse_input function and after we free the error because else we would get leaks.