heredoc.c added
Had to expand heredoc. Becuase it wouldnt expand things and with 'EOF' It would after i added it. Had to make new changes. Then in execute command i made a check if we have a command. Because it tried to run <<EOF but there is no command and we got a segfault.