Export fix
Rearranged some functions to have space. Fixed tabs in minishell.h. Fixed too long line in expand_filename. Fixed export because we checked for a value but we didnt have to, so now it shows export b= in env if we do it once, before we had to do it twice